# Use `new_data_frame()` in `dplyr_reconstruct()` to avoid materializing row names